+++ Changelog.md
@@ -1,5 +1,6 @@
 -----------Changelog for v1.1.3-----------
 
+- Added config to control whether other players should receive join/leave messages when a player (un-)vanishes
 - Fixed that Command Blocks are not able to execute /v by default
 
 -----------Changelog for v1.1.2-----------