+++ Changelog.md
@@ -1,6 +1,7 @@
 -----------Changelog for v1.1.3-----------
 
 - Added config to control whether other players should receive join/leave messages when a player (un-)vanishes
+- Added configs to change some messages of this mod, for more info please check out the mod's config file
 - Fixed that Command Blocks are not able to execute /v by default
 
 -----------Changelog for v1.1.2-----------